requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.7
pathlib==1.0.1
pillow==10.3.0
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
